user controller add update delete get avec les data

+++ app/Http/Controllers/UserController.php
@@ -27,9 +27,9 @@ class UserController extends Controller
      * @param  Request  $request
      * @return Response
      */
-    public function allUsers(Request $request)
+    public function getUsers(Request $request)
     {
-        return response()->json(['users' =>  User::all(), 'usersData' => UserData::all()], 200);
+        return response()->json(['users' =>  User::all()], 200);
     }
     /**
      * Get one user
@@ -37,12 +37,11 @@ class UserController extends Controller
      * @param  Request  $request
      * @return Response
      */
-    public function oneUser($id)
+    public function getUser($id)
     {
         try {
             $user = User::all()->where('idUser', $id)->first();
-            $userData = UserData::all()->where('idUser', $id)->first();
-            return response()->json(['user' => $user, 'userData' => $userData], 200);
+            return response()->json(['user' => $user], 200);
         } catch (\Exception $e) {
 
             return response()->json(['message' => 'User not found!' . $e->getMessage()], 404);
@@ -54,7 +53,7 @@ class UserController extends Controller
      * @param  Request  $request
      * @return Response
      */
-    public function register(Request $request)
+    public function addUser(Request $request)
     {
         //validate incoming request
         $this->validate($request, [
@@ -63,10 +62,9 @@ class UserController extends Controller
             'emailUser' => 'required|email|unique:users',
             'passwordUser' => 'required|confirmed',
             'idRoleUser' => 'required|integer',
-            'keyUserData' => 'string',
-            'valueUserData' => 'string',
             'created_by' => 'required|integer',
             'updated_by' => 'required|integer',
+            'data' => 'string',
         ]);
 
         try {
@@ -79,22 +77,23 @@ class UserController extends Controller
             $user->idRoleUser = $request->input('idRoleUser');
             $user->created_by = $request->input('created_by');
             $user->updated_by = $request->input('updated_by');
-            if (!$user->save())
-                return response()->json(['message' => 'User Registration Failed!'], 409);
 
-            $userData = new UserData;
-            $userData->keyUserData = $request->input('keyUserData') !== null ? $request->input('keyUserData') : '';
-            $userData->valueUserData = $request->input('valueUserData') !== null ? $request->input('valueUserData') : '';
-            $userData->idUser = $user->idUser;
-            $userData->created_by = $request->input('created_by');
-            $userData->updated_by = $request->input('updated_by');
-            $userData->save();
+            $user->save();
+
+            if ($request->input('data') !== null) {
+                $data = (array)json_decode($request->input('data'), true);
+
+                foreach ($data as $key => $value) {
+                    if (!$this->addData($user->idUser, $key, $value, $request))
+                        return response()->json(['message' => 'User data not added!', 'status' => 'fail'], 500);
+                }
+            }
 
             //return successful response
-            return response()->json(['user' => $user, 'userData' => $userData, 'message' => 'CREATED'], 201);
+            return response()->json(['user' => $user, 'message' => 'CREATED', 'status' => 'success'], 201);
         } catch (\Exception $e) {
             //return error message
-            return response()->json(['message' => 'User Registration Failed!' . $e->getMessage()], 409);
+            return response()->json(['message' => 'User Registration Failed!', 'status' => 'fail'], 409);
         }
     }
 
@@ -105,60 +104,7 @@ class UserController extends Controller
      * @param  Request  $request
      * @return Response
      */
-    public function put($id, Request $request)
-    {
-        //validate incoming request
-        $this->validate($request, [
-            'lastnameUser' => 'required|string',
-            'firstnameUser' => 'required|string',
-            'emailUser' => 'required|email|unique:users,emailUser,' . $request->id . ',idUser',
-            'passwordUser' => 'required|confirmed',
-            'keyUserData' => 'required|string',
-            'valueUserData' => 'required|string',
-            'idRoleUser' => 'required|integer',
-            'created_by' => 'required|integer',
-            'updated_by' => 'required|integer',
-        ]);
-
-        try {
-            $user = User::findOrFail($id);
-            $user->lastnameUser = $request->input('lastnameUser');
-            $user->firstnameUser = $request->input('firstnameUser');
-            $user->emailUser = $request->input('emailUser');
-            $plainPassword = $request->input('passwordUser');
-            $user->passwordUser = app('hash')->make($plainPassword);
-            $user->idRoleUser = $request->input('idRoleUser');
-            $user->created_by = $request->input('created_by');
-            $user->updated_by = $request->input('updated_by');
-
-            if (!$user->update())
-                return response()->json(['message' => 'User Update Failed!', 'status' => 'fail'], 409);
-
-            $userData =  UserData::all()->where('idUser', $id)->first();
-            $userData->keyUserData = $request->input('keyUserData');
-            $userData->valueUserData = $request->input('valueUserData');
-            $userData->idUser = $user->idUser;
-            $userData->created_by = $request->input('created_by');
-            $userData->updated_by = $request->input('updated_by');
-            $userData->update();
-
-
-            //return successful response
-            return response()->json(['user' => $user, 'userData' => $userData, 'message' => 'ALL UPDATED', 'status' => 'success'], 200);
-        } catch (\Exception $e) {
-            //return error message
-            return response()->json(['message' => 'User Update Failed!' . $e->getMessage(), 'status' => 'fail'], 409);
-        }
-    }
-
-    /**
-     * Update user patch.
-     *
-     * @param  string   $id
-     * @param  Request  $request
-     * @return Response
-     */
-    public function patch($id, Request $request)
+    public function updateUser($id, Request $request)
     {
         //validate incoming request
         $this->validate($request, [
@@ -166,19 +112,16 @@ class UserController extends Controller
             'firstnameUser' => 'string',
             'emailUser' => 'email|unique:users,emailUser,' . $request->id . ',idUser',
             'passwordUser' => 'confirmed',
-            'keyUserData' => 'string',
-            'valueUserData' => 'string',
             'idRoleUser' => 'integer',
             'created_by' => 'integer',
             'updated_by' => 'integer',
+            
+            'data' => 'string',
         ]);
 
         try {
+            // On modifie les infos principales du user
             $user = User::findOrFail($id);
-
-            if (in_array(null or '', $request->all()))
-                return response()->json(['message' => 'Null or empty value', 'status' => 'fail'], 500);
-
             if ($request->input('lastnameUser') !== null)
                 $user->lastnameUser = $request->input('lastnameUser');
             if ($request->input('firstnameUser') !== null)
@@ -196,23 +139,20 @@ class UserController extends Controller
             if ($request->input('updated_by') !== null)
                 $user->updated_by = $request->input('updated_by');
 
-            if (!$user->update())
-                return response()->json(['message' => 'User Update Failed!', 'status' => 'fail'], 409);
+            $user->update();
 
-            $userData = UserData::all()->where('idUser', $id)->first();
-            if ($request->input('keyUserData') !== null)
-                $userData->keyUserData = $request->input('keyUserData');
-            if ($request->input('valueUserData') !== null)
-                $userData->valueUserData = $request->input('valueUserData');
-            $userData->idUser = $user->idUser;
-            if ($request->input('created_by') !== null)
-                $userData->created_by = $request->input('created_by');
-            if ($request->input('updated_by') !== null)
-                $userData->updated_by = $request->input('updated_by');
-            $userData->update();
+            //maj des data
+            if ($request->input('data') !== null) {
+                $data = (array)json_decode($request->input('data'), true);
+
+                foreach ($data as $key => $value) {
+                    if (!$this->updateData($user->idUser, $key, $value))
+                        return response()->json(['message' => 'User Update Failed!', 'status' => 'fail'], 500);
+                }
+            }
 
             //return successful response
-            return response()->json(['user' => $user, 'userData' => $userData, 'message' => 'PATCHED', 'status' => 'success'], 200);
+            return response()->json(['user' => $user, 'data' => $this->getAllData($user->idUser)->original, 'message' => 'ALL UPDATED', 'status' => 'success'], 200);
         } catch (\Exception $e) {
             //return error message
             return response()->json(['message' => 'User Update Failed!' . $e->getMessage(), 'status' => 'fail'], 409);
@@ -225,18 +165,89 @@ class UserController extends Controller
      * @param int $id
      * @return Response
      */
-    public function delete($id)
+    public function deleteUser($id)
     {
         try {
             $user = User::findOrFail($id);
-            $userData =  UserData::all()->where('idUser', $id)->first();
-            $userData->delete();
+            $userData = UserData::all()->where('idUser', $id);
+
+            //maj des data
+            if ($userData !== null) {
+                foreach ($userData as $key => $value) {
+                    if (!$this->deleteData($user->idUser, $key))
+                        return response()->json(['message' => 'User Deletion Failed!', 'status' => 'fail'], 500);
+                }
+            }
+
             $user->delete();
 
-            return response()->json(['user' => $user, 'userData' => $userData, 'message' => 'DELETED', 'status' => 'success'], 200);
+            return response()->json(['user' => $user, 'data' => $userData, 'message' => 'DELETED', 'status' => 'success'], 200);
         } catch (\Exception $e) {
             //return error message
             return response()->json(['message' => 'User deletion failed!' . $e->getMessage(), 'status' => 'fail'], 409);
+        }
+    }
+
+    public function addData($idUser, $key, $value, $request)
+    {
+        try {
+            $userData = new UserData;
+            $userData->keyUserData = $key;
+            $userData->valueUserData = $value;
+            $userData->created_by = $request->input('created_by');
+            $userData->updated_by = $request->input('updated_by');
+            $userData->idUser = $idUser;
+
+            $userData->save();
+
+            //return successful response
+            return response()->json(['user' => $userData, 'message' => 'CREATED'], 201);
+        } catch (\Exception $e) {
+            //return error message
+            return response()->json(['message' => 'User data not added!' . $e->getMessage()], 409);
+        }
+    }
+
+    public function getAllData($idUser)
+    {
+        return response()->json(UserData::all()->where('idUser', $idUser), 200);
+    }
+
+    public function getData($idUser, $key)
+    {
+        return response()->json(UserData::all()->where('idUser', $idUser)->where('keyUserData', $key), 200);
+    }
+
+    public function updateData($idUser, $key, $value)
+    {
+        try {
+            $userData = UserData::all()->where('idUser', $idUser)->where('keyUserData', $key)->first();
+
+            if ($userData == null)
+                return false;
+
+            $userData->valueUserData = $value;
+            $userData->update();
+
+            return true;
+        } catch (\Exception $e) {
+            return false;
+        }
+    }
+
+    public function deleteData($idUser, $key)
+    {
+        try {
+            $userData = UserData::all()->where('idUser', $idUser)->where('keyUserData', $key)->first();
+
+            if ($userData == null)
+                return false;
+
+            $userData->delete();
+
+            return true;
+        } catch (\Exception $e) {
+            return false;
         }
     }
